+++ R/ArraySpeciesBySize-class.R
@@ -534,6 +534,7 @@ plot2.ArraySpeciesBySize <- function(x, y, name1 = "First", name2 = "Second",
     compare_array_metadata(x, y)
     size_axis <- plot_size_axis(size_axis)
     check_per_log_size(x, per_log_size)
+    log_y <- array_log_y(x, log_y, log, !missing(log_y))
     log_axes <- parsePlotLog(log, log_x = log_x, log_y = log_y)
     log_x <- log_axes$log_x
     log_y <- log_axes$log_y
@@ -550,6 +551,8 @@ plot2.ArraySpeciesBySize <- function(x, y, name1 = "First", name2 = "Second",
     plot_dat2 <- prepare_ArraySpeciesBySize_plot_data(
         y, species = species, all.sizes = all.sizes, wlim = wlim,
         total = total, background = background)
+
+    ylim <- array_ylim(x, ylim, log_y, c(plot_dat1[[2]], plot_dat2[[2]]))
 
     plotComparisonDataFrame(plot_dat1, plot_dat2, params,
                             name1 = name1, name2 = name2,